test(consensus): split restarted node catch-up tests
